requirements.txt
anthropic>=0.39.0
rich>=13.0.0
google-generativeai>=0.5.0
